Working copy of the server capable of calculating and returning absorbance.
git-svn-id: https://cotprojects.marine.usf.edu/svn/MiniSEAS/Gumstix/Source@41 63cbca5c-2ead-4e1b-9715-33247cc11c71

+++ USB4000Gum/source/absorbanceCalculation.h
@@ -13,7 +13,7 @@
 #include "USB4000Gum.h"
 
 float ComputeAbsorbance(spectrometer* USB4000, unsigned short absorbingPixel, unsigned short nonAbsorbingPixel);
-float ComputeCorrectionAbsorbance(spectrometer* USB4000, unsigned int nonAbsorbingPixel);
+float ComputeCorrectionAbsorbance(spectrometer* USB4000, unsigned short nonAbsorbingPixel);
 float ComputeDarkValue(spectrometer* USB4000);
 float NonLinearCountCorrection(spectrometer* USB4000, float Counts);
 void  ComputeSpectrometerLambdaValues(spectrometer* USB4000);
